+++ Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.11.0
+- Change the way file names are handled for models
+- Overhaul the design of the header to save space
+- Move the DirectML toggle from Settings to the header
+- Change the default model type from `tspan` to `image` to reduce chance of error for models without metadata
+
 ## 0.10.2
 - Implement segment selection. Users can now select a small segment of a video to process and preview!
   - When using this mode, the comparison buttons are disabled